Revert "Revert "Further pom corrections""
This reverts commit f06a6d68323133e8ee602cfd70d7c54eec3b15c6.

+++ src/main/java/edu/southwestern/util/graphics/GraphicsUtil.java
@@ -389,10 +389,10 @@ public class GraphicsUtil {
 	 * @return BufferedImage 
 	 */
 	public static BufferedImage imageFromINDArray(INDArray array) {
-		int[] shape = array.shape();
+		long[] shape = array.shape();
 		// Should the order of these be switched?
-		int width = shape[2];
-		int height = shape[3];
+		int width = (int) shape[2];
+		int height = (int) shape[3];
 		BufferedImage image = new BufferedImage(width, height, BufferedImage.TYPE_INT_RGB);
 		// Copy from INDArray to BufferedImage
 		for (int x = 0; x < width; x++) {// scans across whole image
